Latest

+++ linux/fetch/gen_sh.py
@@ -1,5 +1,7 @@
 import os
 
 def gen_sh(path, lines):
+  if os.path.isfile(path):
+    return
   open(path, "w").write("\n".join(lines))
   os.chmod(path, 0x755)
